+++ rocNet.Kernel/Routing/RouteEntry.cs
@@ -5,7 +5,7 @@ namespace LibKernel
     public interface RouteEntry
     {
         Guid GroupId { get; }
-        int Energy { get; }
+        long Energy { get; }
         bool Match(string nri);
         Func<ResourceRequest, ResourceRepresentation> Handler { get; }
     }
